requirements.txt
beautifulsoup4>=4.9.0
requests>=2.25.0
python-dateutil>=2.8.0
openpyxl>=3.0.0  
PyYAML>=5.4.0
schedule>=1.1.0